Separate map data from app
Introduces reading map data from sdcard.

- usage of device storage api
- introduction of the functional tile layers plugin for leaflet in order to access that api
- update of leaflet to a functional tile layers supported version (0.7.7 seems to work as well as 0.7.2)
- the mid term goal is to leave a simple world map with inline data as default map and select more maps in the menu

Usage of the device storage api requires switching from web app type to a priviledged app. This is pain, because this app is _never_ going to marketplace, that would not make sense for a packaged app in most cases, but definitely not in this one. Anyway, this currently means the app can probably only be installed via the web IDE. This is fine for me (it is the easiest deployment for me anyway), but I still feels unsatisfied. There is however no way that I know of to avoid that. Storing the map data in the app is no long term solution because of limited storage space on the system drive and recopying that data on deployment is unnecessary and unstable. This way uploading additional map data can be done without touching the app.